Update Main_Control_Unit_Testing.ino
Got rid of the `|` in places where it'll be hard to parse, replaced it with spaces instead. Got rid of the `debug` boolean flag and replaced it with `dashboard_btn_testing` and `pedal_input_testing` wherever necessary. Added a new boolean variable `torque_debugging` to take care of the parts of the code dealing with code that can be used for debugging.